pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wdim-eval"
version = "0.1.0"
description = "WDIM evaluation playground for Apple Foundation Models"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "ffmpeg-python",
    "jupyter",
    "ipywidgets",
    "tiktoken",
    "apple-fm-sdk",
    "rich",
    "pandas",
    "papermill",
]

[tool.setuptools.packages.find]
include = ["lib*"]
